Move evidence_fields to shared package root
