+++ mcp_fuzzer/reports/formatters/common.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Any, Iterable, Literal, Protocol
 
-from ..evidence_fields import (
+from ...evidence_fields import (
     ACCEPTED_MALFORMED,
     ERROR,
     EXCEPTION,
